requirements.txt
streamlit==1.35.0
requests==2.31.0
pandas>=2.2.3
plotly==5.22.0
python-dotenv==1.0.1
schedule==1.2.2
fpdf2==2.7.9